Moved run_opt into test file.
